* rework some data structures & date sort
- Change the data structure that contains the output to use a dictionary
so that I can rearrange the headers

- Try to parse the DATE: value in the header.  Its suprisingly hard to find
a module that will do that.  The current one almost works but it chokes on
logfiles from Peru that have the Day in spanish. (ie Lun for Mon).  I have to go
delete that from the string.
TODO: Go change the DATE: string in the logfile to use a more structured
date format.