Changes for release to Quanta for XO-4
There are many changes in this change set most of them changes to
thresholds while trying debug the laptop power off thats happening on
XO-4 sometimes when runin dischage is disabled.

First try at setting the XO-4 audio output so that the power draw of
the test matches that of the XO-1.75

A functional change is to try and detect when the EC turns off
runin-discharge before this script detects that it needs to turn it
off.  This prevents the script from never exiting because the low
voltage threshold was never reached.